Game Version 0.1.5.5
Updated SRP to 1.11.1. Platforms 0.8.0p3
Butcher is now interactable.
Butcher can now be talked with, and interacted through his shop.
Created a unified NPC Inventory System.
Fixed a minor bug with interactable objects.
Fixed a bug with zoomed-in camera on how it handles occluded views.
Added new items in the game, that are sold by the butcher.
+ Chicken, Beef, Pork, Lamb Items.