Game Version 0.1.6.0
Can now buy drinks, up to 5 at a time.
Minor change to sort algorithm for potential performance increase.
Adjusted drink prices to reflect their price.
If the player cannot buy 5 drinks, then the Shop will show the maximum amount the player can buy up to 5.
Pricing of the drinks reflects the correct price that the player will pay.
Fixed a bug with inventory UI not updating when purchasing items that are directly added to inn.
Updated incurring costs to include newly purchased assets, will be included in project at a later version.
Fixed if the merchant does not have enough drinks to sell 5 at a time, then the merchant will sell up to however many they have left.
Lowered the maximum amount of drinks that can be stored at the in..