Game Version 0.1.6.5
Project Upgraded to Unity 2020.1.5.

Roaster now rotates if there is any recipes on it.
Increased the time required to hold to rotate the player camera. (0.2 seconds -> 0.3 seconds)
Fixed some areas of the grill couldn't be right clicked.
Added Bubbling Effect while cooking on the stove.
Added an emptying effect if the food is burnt or undercooked while using the stove.
Increased the visibility of Heat Distortion Effect(0.07 -> 0.082) to make it more inline with other effects.
Removed unnecessary LuxURP folders.