Game Version 0.1.7.6
Drunk is added to the ecosystem with limited functionality.
Drunk can be found at the mill's dock between 3:30 - 12:00.
Every other time, he'll be wandering around town and drinking.

Woodworker is added to the ecosystem.
Woodworker will be working between 9:30 - 13:30 and 14:30 - 18:00.
Woodworker will be at home between 22:00 - 9:00.
Woodworker will be wandering the world every other time.

Miller is added to the ecosystem.
Miller will be working between 6:30 - 12:00 and 14:00 - 16:30.
Miller will be at home between 20:50 - 6:30.
Miller can be found around the town every other time.

Fixed not being able to remove other agents from the world.
Fixed if an agent was not interactable, it would cause an exception when trying to remove them from the world.
Fixed a crash caused by agents searching for a path, before AStar was initialized.